# frontend/client-code-gen/fam-openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Forest Access Management - FAM - API",
    "description" : "Forest Access Management API. Defines who has access to which applications, and the roles they operate under once access is granted.\n\nCovers both the app-access-control and admin-management surfaces that upstream FAM served from two separate APIs.",
    "contact" : {
      "name" : "Team Heartwood",
      "url" : "https://apps.nrs.gov.bc.ca/int/confluence/display/FSAST1/Team+Heartwood",
      "email" : "SIBIFSAF@Victoria1.gov.bc.ca"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Relative to the host serving this document"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/user-role-assignment" : {
      "post" : {
        "tags" : [ "FAM User Role Assignment" ],
        "summary" : "Grant multiple users access to an application's role",
        "description" : "Granting IDIR/BCeID users access to an application's role, supporting expiry dates for role assignments.",
        "operationId" : "create_user_role_assignment_many",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FamUserRoleAssignmentCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FamUserRoleAssignmentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/fam-applications/{applicationId}/users/{userId}" : {
      "get" : {
        "tags" : [ "FAM Applications" ],
        "summary" : "Retrieve user information by user ID under an application",
        "operationId" : "get_application_user_by_id",
        "parameters" : [ {
          "name" : "applicationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FamUserInfoDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/user-role-assignment/{userRoleXrefId}" : {
      "delete" : {
        "tags" : [ "FAM User Role Assignment" ],
        "summary" : "Remove a specific application's role from a user's access",
        "operationId" : "delete_user_role_assignment",
        "parameters" : [ {
          "name" : "userRoleXrefId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "FamUserRoleAssignmentCreateRequest" : {
        "properties" : {
          "users" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/FamUserRoleAssignmentUserDto"
            },
            "maxItems" : 50,
            "minItems" : 0
          },
          "user_type_code" : {
            "$ref" : "#/components/schemas/UserType"
          },
          "role_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "forest_client_numbers" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "maxLength" : 8,
              "minLength" : 1
            }
          },
          "requires_send_user_email" : {
            "type" : "boolean"
          },
          "expiry_date_date" : {
            "type" : "string",
            "description" : "The expiry date as a string (YYYY-MM-DD), BC timezone. If provided, the role is valid until the end of this day."
          }
        },
        "required" : [ "role_id", "user_type_code", "users" ]
      },
      "FamUserRoleAssignmentUserDto" : {
        "properties" : {
          "user_name" : {
            "type" : "string",
            "maxLength" : 20,
            "minLength" : 3
          },
          "user_guid" : {
            "type" : "string",
            "maxLength" : 32,
            "minLength" : 32
          }
        },
        "required" : [ "user_guid", "user_name" ]
      },
      "UserType" : {
        "type" : "string",
        "enum" : [ "I", "B", "CD", "CT", "CP" ]
      },
      "EmailSendingStatus" : {
        "type" : "string",
        "enum" : [ "NOT_REQUIRED", "SENT_TO_EMAIL_SERVICE_SUCCESS", "SENT_TO_EMAIL_SERVICE_FAILURE" ]
      },
      "FamApplicationDto" : {
        "properties" : {
          "application_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "application_name" : {
            "type" : "string"
          },
          "application_description" : {
            "type" : "string"
          }
        },
        "required" : [ "application_description", "application_id", "application_name" ]
      },
      "FamApplicationUserRoleAssignmentGetDto" : {
        "properties" : {
          "user_role_xref_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "user_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "role_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "user" : {
            "$ref" : "#/components/schemas/FamUserInfoDto"
          },
          "role" : {
            "$ref" : "#/components/schemas/FamRoleWithClientDto"
          },
          "create_date" : {
            "type" : "string",
            "format" : "date-time"
          },
          "expiry_date" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "create_date", "role", "role_id", "user", "user_id", "user_role_xref_id" ]
      },
      "FamForestClientDto" : {
        "properties" : {
          "client_name" : {
            "type" : "string"
          },
          "forest_client_number" : {
            "type" : "string"
          },
          "status" : {
            "$ref" : "#/components/schemas/FamForestClientStatusDto"
          }
        },
        "required" : [ "forest_client_number" ]
      },
      "FamForestClientStatusDto" : {
        "properties" : {
          "status_code" : {
            "$ref" : "#/components/schemas/FamForestClientStatusType"
          },
          "description" : {
            "type" : "string"
          }
        }
      },
      "FamForestClientStatusType" : {
        "type" : "string",
        "enum" : [ "A", "I" ]
      },
      "FamRoleMinDto" : {
        "properties" : {
          "role_name" : {
            "type" : "string"
          },
          "role_type_code" : {
            "type" : "string"
          },
          "application" : {
            "$ref" : "#/components/schemas/FamApplicationDto"
          }
        },
        "required" : [ "application", "role_name", "role_type_code" ]
      },
      "FamRoleWithClientDto" : {
        "properties" : {
          "role_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "role_name" : {
            "type" : "string"
          },
          "role_type_code" : {
            "type" : "string"
          },
          "display_name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "application" : {
            "$ref" : "#/components/schemas/FamApplicationDto"
          },
          "forest_client" : {
            "$ref" : "#/components/schemas/FamForestClientDto"
          },
          "parent_role" : {
            "$ref" : "#/components/schemas/FamRoleMinDto"
          }
        },
        "required" : [ "application", "description", "forest_client", "role_id", "role_name", "role_type_code" ]
      },
      "FamUserInfoDto" : {
        "properties" : {
          "user_name" : {
            "type" : "string"
          },
          "user_type" : {
            "$ref" : "#/components/schemas/FamUserTypeDto"
          },
          "first_name" : {
            "type" : "string"
          },
          "last_name" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          }
        },
        "required" : [ "user_name", "user_type" ]
      },
      "FamUserRoleAssignmentCreateResponse" : {
        "properties" : {
          "status_code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "detail" : {
            "$ref" : "#/components/schemas/FamApplicationUserRoleAssignmentGetDto"
          },
          "error_message" : {
            "type" : "string"
          },
          "email_sending_status" : {
            "$ref" : "#/components/schemas/EmailSendingStatus"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "required" : [ "status_code" ]
      },
      "FamUserRoleAssignmentResponse" : {
        "properties" : {
          "assignments_detail" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/FamUserRoleAssignmentCreateResponse"
            }
          }
        },
        "required" : [ "assignments_detail" ]
      },
      "FamUserTypeDto" : {
        "properties" : {
          "code" : {
            "$ref" : "#/components/schemas/UserType"
          },
          "description" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "description" ]
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "Access token issued by BC Gov SSO (Keycloak).",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "apiKey" : {
        "type" : "apiKey",
        "description" : "Shared secret for the scheduled user-information refresh.",
        "name" : "X-API-Key",
        "in" : "header"
      }
    }
  }
}